+++ pyannote/core/matrix.py
@@ -65,6 +65,7 @@ class LabelMatrix(object):
                 val = self.df.at[row, col]
                 if not np.isnan(val):
                     yield row, col, val
+
     def iter_values(self):
         for row in self.get_rows():
             for col in self.get_columns():
@@ -181,22 +182,28 @@ class LabelMatrix(object):
     def __str__(self):
         return str(self.df)
 
+    def _repr_html_(self):
+        return self.df._repr_html_()
+
+
 def get_cooccurrence_matrix(R, C):
 
+    # initialize label matrix with zeros
     rows = R.labels()
     cols = C.labels()
-    nRows = len(rows)
-    nCols = len(cols)
+    K = np.zeros((len(rows), len(cols)), dtype=np.float)
+    M = LabelMatrix(data=K, rows=rows, columns=cols)
+    
+    # loop on intersecting tracks
+    for (r_segment, r_track), (c_segment, c_track) in R.co_iter(C):
+        # increment 
+        r_label = R[r_segment, r_track]
+        c_label = C[c_segment, c_track]
+        duration = (r_segment & c_segment).duration
+        M[r_label, c_label] += duration
 
-    K = np.zeros((nRows, nCols), dtype=np.float)
-    for r, row in enumerate(rows):
-        row_coverage = R.label_coverage(row)
-        for c, col in enumerate(cols):
-            col_coverage = C.label_coverage(col)
-            coverage = row_coverage.crop(col_coverage, mode='intersection')
-            K[r, c] = coverage.duration()
+    return M
 
-    return LabelMatrix(data=K, rows=rows, columns=cols)
 
 
 def get_tfidf_matrix(words, documents, idf=True, log=False):