+++ pyannote/core/transcription.py
@@ -12,8 +12,8 @@
 # copies of the Software, and to permit persons to whom the Software is
 # furnished to do so, subject to the following conditions:
 
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
+# The above copyright notice and this permission notice shall be included in
+# all copies or substantial portions of the Software.
 
 # THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 # IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
@@ -31,8 +31,7 @@ from networkx.readwrite.json_graph import node_link_data, node_link_graph
 from time import T, TStart, TEnd
 from segment import Segment
 from json import PYANNOTE_JSON_TRANSCRIPTION
-
-import itertools
+from util import pairwise
 
 
 class Transcription(nx.MultiDiGraph):
@@ -71,7 +70,8 @@ class Transcription(nx.MultiDiGraph):
         if t1.anchored and t2.anchored:
             assert t1 <= t2
 
-        super(Transcription, self).add_edge(t1, t2, key=key, attr_dict=attr_dict, **attrs)
+        super(Transcription, self).add_edge(
+            t1, t2, key=key, attr_dict=attr_dict, **attrs)
 
     def relabel_drifting_nodes(self, mapping=None):
         """Relabel drifting nodes
@@ -80,7 +80,6 @@ class Transcription(nx.MultiDiGraph):
         ----------
         mapping : dict, optional
             A dictionary with the old labels as keys and new labels as values.
-            
 
         Returns
         -------
@@ -104,7 +103,7 @@ class Transcription(nx.MultiDiGraph):
 
         Parameters
         ----------
-        source : Segment, 
+        source : Segment
         target : float or str, optional
 
         Returns
@@ -112,7 +111,7 @@ class Transcription(nx.MultiDiGraph):
         g : Transcription
             Sub-graph between source and target
         """
-        
+
         if isinstance(source, Segment):
             source, target = source.start, source.end
 
@@ -122,7 +121,7 @@ class Transcription(nx.MultiDiGraph):
         if source.anchored:
             before = [n for n in self.anchored() if n <= source]
             if before:
-                source = sorted(before)[-1] 
+                source = sorted(before)[-1]
 
         if target.anchored:
             after = [n for n in self.anchored() if n >= target]
@@ -147,7 +146,7 @@ class Transcription(nx.MultiDiGraph):
         ----------
         drifting_t :
             Existing drifting time in graph
-        another_t : 
+        another_t :
             Existing time in graph
         """
         # drifting_t and another_t must exist in graph
@@ -170,7 +169,7 @@ class Transcription(nx.MultiDiGraph):
     def anchor(self, drifting_t, anchored_t):
         """
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        
+
         o -- [ D ] -- o  ==>  o -- [ A ] -- o
 
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
@@ -179,9 +178,9 @@ class Transcription(nx.MultiDiGraph):
 
         Parameters
         ----------
-        drifting_t : 
+        drifting_t :
             Drifting time to anchor
-        anchored_t : 
+        anchored_t :
             When to anchor `drifting_t`
 
         """
@@ -202,13 +201,13 @@ class Transcription(nx.MultiDiGraph):
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
         o -- [ F ] -- o      o          o
-                               ⟍     ⟋   
+                               ⟍     ⟋
                         ==>     [ F ]
                                ⟋     ⟍
-        o -- [ f ] -- o      o          o    
+        o -- [ f ] -- o      o          o
 
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
-        
+
         Align two (potentially drifting) times
 
         `one_t` and `another_t` cannot both be anchored at the same time
@@ -246,10 +245,10 @@ class Transcription(nx.MultiDiGraph):
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
         p -- [ t1 ]       p         [ t1 ]
-                            ⟍     ⟋   
+                            ⟍     ⟋
                      ==>     [ t ]
                             ⟋     ⟍
-        p' -- [ t2 ]      p'        [ t2 ]    
+        p' -- [ t2 ]      p'        [ t2 ]
 
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
@@ -293,10 +292,10 @@ class Transcription(nx.MultiDiGraph):
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
 
         [ t1 ] -- s       [ t1 ]         s
-                                ⟍     ⟋   
+                                ⟍     ⟋
                      ==>         [ t ]
                                 ⟋     ⟍
-        [ t2 ] -- s'      [ t2 ]        s'    
+        [ t2 ] -- s'      [ t2 ]        s'
 
         ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
         """
@@ -367,14 +366,47 @@ class Transcription(nx.MultiDiGraph):
 
         return g
 
-    def ordered_edges_iter(self, data=False, keys=False):
-        """Return an iterator over the edges in temporal/topological order.
+    def temporal_sort(self):
+        """Get nodes sorted in temporal order
+
+        Remark
+        ------
+        This relies on a combination of temporal ordering of anchored times
+        and topological ordering for drifting times.
+        To be 100% sure that one drifting time happens before another time,
+        check the ordering graph (method .ordering_graph()).
+        """
+
+        g = nx.DiGraph()
+
+        # add times
+        for t in self.nodes_iter():
+            g.add_node(t)
+
+        # add existing edges
+        for t1, t2 in self.edges_iter():
+            g.add_edge(t1, t2)
+
+        # connect pairs of consecutive anchored times
+        anchored = sorted(self.anchored())
+        for t1, t2 in pairwise(anchored):
+            g.add_edge(t1, t2)
+
+        return nx.topological_sort(g)
+
+    # =========================================================================
+
+    def ordered_edges_iter(self, nbunch=None, data=False, keys=False):
+        """Return an iterator over the edges in temporal order.
 
         Ordered edges are returned as tuples with optional data and keys
         in the order (t1, t2, key, data).
 
         Parameters
         ----------
+        nbunch : iterable container, optional (default= all nodes)
+            A container of nodes. The container will be iterated
+            through once.
         data : bool, optional (default=False)
             If True, return edge attribute dict with each edge.
         keys : bool, optional (default=False)
@@ -384,79 +416,89 @@ class Transcription(nx.MultiDiGraph):
         -------
         edge_iter : iterator
             An iterator of (u,v), (u,v,d) or (u,v,key,d) tuples of edges.
+
+        Notes
+        -----
+        Nodes in nbunch that are not in the graph will be (quietly) ignored.
+        For the same reason you should not completely trust temporal_sort,
+        use ordered_edges_iter with care.
         """
 
-        # start by sorting nodes in temporal+topological order
-        o = self.ordering_graph()
-        nodes = nx.topological_sort(o)
+        # start by sorting nodes in temporal order
+        nodes = self.temporal_sort()
 
-        # iterate over edges using this very order
-        for _ in self.edges_iter(nbunch=nodes, data=data, keys=keys):
-            yield _
+        # only keep nbunch subset (while preserving the order)
+        if nbunch:
+            nbunch = list(nbunch)
+            nodes = [n for n in nodes if n in nbunch]
+
+        # iterate over edges using temporal order
+        return self.edges_iter(nbunch=nodes, data=data, keys=keys)
 
     # =========================================================================
 
-    def _anchored_successors(self, n):
-        """Get all first anchored successors"""
+    def timerange(self, t1, t2, inside=True, sort=None):
+        """Infer edge timerange from graph structure
 
-        # loop on all outgoing edges
-        for t in self.successors(n):
-            
-            # if neighbor is anchored
-            # stop looking for (necessarily later) successors
-            if t.anchored:
-                yield t
-                continue
+        a -- ... -- [ t1 ] -- A -- ... -- B -- [ t2 ] -- ... -- b
 
-            # if neighbor is not anchored
-            # look one level deeper
-            for tt in self._anchored_successors(t):
-                yield tt
+        ==> [a, b] (inside=False) or [A, B] (inside=True)
 
-    def _anchored_predecessors(self, n):
-        """Get all first anchored predecessors"""
-
-        # loop on all incoming edges
-        for t in self.predecessors(n):
-            
-            # if predecessor is anchored
-            # stop looking for (necessarily earlier) predecessors
-            if t.anchored:
-                yield t
-                continue
-            
-            # if neighbor is not anchored
-            # look one level deeper
-            for tt in self._anchored_predecessors(t):
-                yield tt
-
-    def timerange(self, t):
-        """Infer smallest possible timerange from graph structure
+        Parameters
+        ----------
+        t1, t2 : anchored or drifting times
+        inside : boolean, optional
 
         Returns
         -------
-        (left, right) tuple
-            left == None or right == None indicates that the current state of
-            the annotation graph does not allow to decide the boundary.
-
+        segment : Segment
         """
 
-        t = T(t)
+        t1 = T(t1)
+        t2 = T(t2)
 
-        if t.anchored:
-            return (t.T, t.T)
-        successors = [n for n in self._anchored_successors(t)]
-        predecessors = [n for n in self._anchored_predecessors(t)]
+        # in case it is not provided, compute temporal sort
+        if sort is None:
+            sort = self.temporal_sort()
 
-        earlier_successor = None
-        if successors:
-            earlier_successor = min(successors)
+        # if edge start is anchored, use it as start time
+        if t1.anchored:
+            start = t1
 
-        later_predecessor = None
-        if predecessors:
-            later_predecessor = max(predecessors)
+        # otherwise, look for the closest anchored time in temporal order:
+        # - just after if inside is True
+        # - just before otherwise
+        else:
+            start = None
+            # find time index in temporal sort
+            istart = sort.index(t1)
+            # search just before or just after depending on 'inside' value
+            search = sort[istart+1:] if inside else sort[istart-1::-1]
+            for t in search:
+                if t.anchored:
+                    start = t
+                    break
+            # if we could not find any anchored time
+            # use document end of start depending on 'inside' value
+            if start is None:
+                start = TEnd if inside else TStart
 
-        return (later_predecessor.T, earlier_successor.T)
+        # same treatment for the other end of edge
+        if t2.anchored:
+            end = t2
+        else:
+            end = None
+            iend = sort.index(t2)
+            search = sort[iend-1::-1] if inside else sort[iend+1:]
+            for t in search:
+                if t.anchored:
+                    end = t
+                    break
+            if end is None:
+                end = TStart if inside else TEnd
+
+        # return a 'Segment'
+        return Segment(start=start, end=end)
 
     # =========================================================================