+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+### Version 0.0.3 (2014-06-02)
+
+  - feat(Annotation):     add 'collar' param to .smooth()
+  - refactor(Annotation): remove support for >> operator
+  - maintain(Mapping):    remove label mapping data structure
+  - feat(LabelMatrix):    add IPython display
+  - improve(LabelMatrix): 10x faster cooccurrence matrix
+  - feat(Scores):         add IPython display
+  - feat(Transcription):  add edge timerange prediction
+  - feat(Transcription):  add node temporal sort
+  - fix(Transcription):   make label_timeline return a copy
+  - fix(Transcription):   fix IPython display
+  - docs(Transcription):  add IPython documentation for Transcription
+
 ### Version 0.0.2 (2014-05-06)
 
   - feat: Transcription data structure (annotation graph)
@@ -5,4 +19,3 @@
 ### Version 0.0.1 (2014-05-02)
 
   - first public version
-